Changed Footprint and Trap into Interactables
