+++ Nightmare Library/Assets/Jack/Prefabs/Evidence/Footprint.prefab
@@ -47,6 +47,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6f0550ce63b6b0840b5b9876cfd08c8f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  placementTypes: 
+  floorPlacementType: 0
+  wallPlacementType: 0
+  fixPlacement: 1
+  precisePlacement: 0
+  allowPlayerClick: 0
+  allowPlayerPickup: 0
+  allowEnemyHysterics: 0
+  allowEnemyFlicker: 0
+  attachedLight: {fileID: 0}
 --- !u!114 &827733343639628691
 MonoBehaviour:
   m_ObjectHideFlags: 0
