+++ Nightmare Library/Assets/Jack/Scriptable Objects/EnemyPresets/Testenemy2.asset
@@ -10,10 +10,10 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 44d86c9be31808c4c83bd98820d6d8cc, type: 3}
-  m_Name: TestEnemy
+  m_Name: Testenemy2
   m_EditorClassIdentifier: 
-  enemyName: Test Enemy
-  description: A Basic Test Enemy
-  activeAttacks: 02000000
+  enemyName: Test Enemy 2
+  description: The second test enemy
+  activeAttacks: 03000000
   passiveAttacks: 000000000100000002000000
-  evidence: 030000000200000001000000
+  evidence: 030000000400000005000000
