+++ Nightmare Library/Assets/Jack/Scriptable Objects/EnemyPresets/Testenemy2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e0beca956d577944acf20ccc0bab9a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
