+++ Nightmare Library/Assets/Jack/_Scripts/Enemy Scripts/Evidence/FootprintNetwork.cs
@@ -4,8 +4,10 @@ using System.Collections.Generic;
 using Unity.Netcode;
 using UnityEngine;
 
+using NetVar;
+
 [RequireComponent(typeof(FootprintController))]
-public class FootprintNetwork : NetworkBehaviour
+public class FootprintNetwork : InteractableNetwork
 {
     private FootprintController fController;
 
@@ -15,10 +17,7 @@ public class FootprintNetwork : NetworkBehaviour
 
         if (IsOwner)
         {
-            fController.OnFootprintSpawn += OnFootprintSpawn;
-            fController.OnFootprintDespawn += OnFootprintDespawn;
-
-            OnFootprintDespawn(this, EventArgs.Empty);
+            OnPickup(true);
         }
         else
         {
@@ -27,34 +26,5 @@ public class FootprintNetwork : NetworkBehaviour
         }
 
         gameObject.SetActive(false);
-    }
-
-    private void OnFootprintSpawn(object sender, EventArgs e)
-    {
-        OnFootprintSpawnClientRpc(transform.position);
-    }
-    private void OnFootprintDespawn(object sender, EventArgs e)
-    {
-        OnFootprintDespawnClientRpc();
-    }
-
-    [ClientRpc]
-    public void OnFootprintSpawnClientRpc(Vector3 pos)
-    {
-        if (!IsOwner)
-        {
-            transform.position = pos;
-            gameObject.SetActive(true);
-
-            fController.Activate();
-        }
-    }
-    [ClientRpc]
-    public void OnFootprintDespawnClientRpc()
-    {
-        if (!IsOwner)
-        {
-            fController.Deactivate();
-        }
     }
 }
