+++ Nightmare Library/Assets/Jack/_Scripts/Helpful Classes/NetworkVariables.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01d651fd2dabe9d449162d4198485757
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
