+++ Nightmare Library/Assets/Jack/_Scripts/Interactables/Monitors/MonitorCameraController.cs
@@ -22,10 +22,10 @@ public class MonitorCameraController : Interactable
         cam.targetTexture = renderTexture;
     }
 
-    public override void Pickup(bool fromNetwork = false)
+    public override GameObject Pickup(bool fromNetwork = false)
     {
         SetBroadcasting(false);
-        base.Pickup(fromNetwork);
+        return base.Pickup(fromNetwork);
     }
     public override void Place(bool fromNetwork = false)
     {
